refactor(image): utility logic change to image ctrl
